moved eeprom store into its own file
